pyproject.toml
[project]
name = "ai-newsletter"
version = "0.1.0"
description = "Personal automated daily intelligence agent - YouTube, blogs, and email"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.100.0",
    "youtube-transcript-api>=0.6.0",
    "anthropic>=0.25.0",
    "requests>=2.32.5",
    "feedparser>=6.0.12",
    "openai>=2.15.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "ipykernel>=7.1.0",
    "pyyaml>=6.0.3",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.46",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "streamlit>=1.53.1",
]
